requirements.txt
streamlit
pandas
numpy
torch
rdkit-pypi
scikit-learn
requests
biopython
shap
plotly
pyyaml
tqdm
openpyxl
joblib
pyarrow
